refactor(wss): split SinglePathOp into OpAdapterMeta + Op<W>; relocate the trial machinery (nbl.18)
Harness hygiene ahead of the WSS PR, behaviour-preserving: the matrix is
byte-identical before and after (796 pass / 296 fail / 301 pending, 0
newly-failing, the same 5 known batch-ahead un-pend candidates).

- SPLIT `SinglePathOp<W>` into a supertrait pair. `OpAdapterMeta` carries the
  op-level, world-agnostic surface — `name`, `cases`, `ok_effect` — stated ONCE
  per op; `Op<W: Layer>: OpAdapterMeta` carries only `invoke`, plus a default
  `ok_effect` delegating to the op-level one so a world can override it when its
  successful effect genuinely differs. Every op previously restated
  name/cases/ok_effect in each of its 3-4 world impls (8x for move's two roles);
  now it is 1x (2x for move). `OpAdapterMeta::ok_effect` is REQUIRED, not
  defaulted: a default `Ok(())` would silently pass an `Ok` cell whose specific
  effect was never checked, and every op already had one.
- MOVE the shared trial machinery out of `adapters/mod.rs` into a new
  `harness/op.rs` — `Case`, the trait pair, `cell_trial`, `run_single_cell`,
  `REL`, `present_state`, and `single_path_trials` (RENAMED `op_trials`; "single
  path" was stale — the mold backs every op including move's MoveSrc/MoveDest).
  `adapters/mod.rs` is now just the submodule declarations.
- The new module is the PORTABLE CORE seam for the later crate extraction
  (turbovault-nbl.13): it is generic over the abstract `Layer`/`Vault` and knows
  nothing about turbovault's substrates, so the extraction becomes a move rather
  than a rewrite. The turbovault glue stays in `backend.rs`.
- HOIST the batch `observe()` to the call site: `BatchWorld::run_batch_of_one`
  (work + wrap, returning `Observed`) becomes `apply_op` (returning the raw
  substrate `Result`), so a batch invoker reads
  `observe(w.apply_op(op).await, w.vault().read(rel))` — the same shape as the
  Tools/Manager arms, instead of the one place a World's mechanism leaked into
  the adapters.
- The `ok_effect` name now exists on both traits, so `run_single_cell`
  disambiguates via the subtrait form (`<A as Op<W>>::ok_effect`) — a world's
  override wins, else it delegates.

Retires the `SinglePathOp` / `single_path_trials` / `run_batch_of_one` names
entirely (comments included). `just test` green.
